refactor gendata for tests
