+++ kernels/simple_matmul/gendata.py
@@ -1,43 +1,8 @@
 # simple script to generate inputs and expected outputs for simple_matmult
 
 import numpy as np
-from numpy import typing as npt
 
-
-def create_header(
-    file_name: str, sizes: dict[str, int], variables: dict[str, npt.NDArray]
-) -> None:
-    with open(file_name, "w") as f:
-        includes = ["#include <stdint.h>", "#pragma once", ""]
-        includes = "\n".join(includes)
-        variables_string = [""]
-        for i, j in sizes.items():
-            variables_string.append(f"#define {i} {j}")
-        variables_string.append("")
-        for i, j in variables.items():
-            variables_string.append(f"extern const {j.dtype}_t {i}[{j.size}];")
-        variables_string = "\n".join(variables_string)
-        f.write(includes)
-        f.write(variables_string)
-        f.write("\n")
-
-
-def create_data(file_name: str, variables: dict[str, npt.NDArray]):
-    includes = ['#include "data.h"', "", ""]
-    includes = "\n".join(includes)
-    variables = {i: np.reshape(j, j.size) for i, j in variables.items()}
-    with open(file_name, "w") as f:
-        f.write(includes)
-        for variable_name, variable_value in variables.items():
-            f.write(
-                f"const {variable_value.dtype}_t {variable_name}"
-                + f"[{variable_value.size}] = "
-                + "{\n"
-            )
-            variable_str = ["\t" + str(i) for i in variable_value]
-            f.write(",\n".join(variable_str))
-            f.write("\n};\n\n")
-
+from util.gendata import create_data, create_header
 
 if __name__ == "__main__":
     # Reset random seed for reproducible behavior
@@ -55,14 +20,17 @@ if __name__ == "__main__":
     C_golden = np.matmul(A.astype(np.dtype("int32")), B.astype(np.dtype("int32")))
     C = np.zeros(C_golden.shape, np.dtype("int32"))
 
-    sizes = {"N_size": A.shape[0], "K_size": A.shape[1], "M_size": B.shape[1]}
-
     # Perform layout transformations before writing to memory
 
     # only thing necessary: transform B from row-major to column-major
     B_new_layout = np.transpose(B)
 
     # C are just all zeros, so layout not important
+    sizes = {
+        "N_size": A.shape[0],
+        "K_size": A.shape[1],
+        "M_size": B.shape[1],
+    }
     variables = {
         "A": A,
         "B": B_new_layout,
